tests/submission/Infraware/Forms/contents/Forms/textarea_selectionend_get.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>input_selectionEnd_get</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<p>
		input: <textarea id='textField'>input</textarea>
	</p>

	<script>

		var textField = document.getElementById("textField");
		textField.select();
		textField.selectionEnd = 3;

		test( function() {
			assert_equals(textField.selectionEnd, 3, "Value of selectionStart attribute is not 3.");
		});

	</script>

 </body>
</html>
